==> scripts/03_imputaciones.qmd <==
---
title: "Imputaciones"
format: docx
editor: visual
---

## 

```{r}
#| include: false
#| label: libraries

library(flextable)
library(moments)
library(tidyverse)
library(tidymodels)
```

```{r}
#| include: false

# list_logs_to_impute <-read_rds("../data/list_logs_to_impute.rds")
for_input <- read_rds("../data/brent_qc3.rds")

list_logs_to_impute <- tibble(Variable_Name = c("CALI","RDEP", "RSHA", "RMED", "SP", "DTC", "NPHI", "DRHO", "rhob_clean", "pef_clean", "GR"  ))
```

### Metodos de imputacion

Considerando que los pozos son prácticamente verticales y los valores nulos para las coordenadas x,y. z vamos a utilizar una estrategia diferente

x y el valor de la ultima coordenada

z, vamos a utilizar el promedio del intervalo de muestra para continuar profundizando el pozo

```{r}
#| label: imputacion_de_coordenadas
#| echo: false


# Estimamos el valor de coord x y y para la ultima coordenada con datos y el intervalo de muetrra en la profundidad
help_coord_imput <- for_input %>% 
  drop_na(x_loc) %>% 
  group_by(well) %>% 
  mutate(depth_step = DEPT -lag(DEPT), 
         depth_step = ifelse(is.na(depth_step), mean(depth_step, na.rm = T), depth_step)) %>% 
  filter(DEPT == max(DEPT)) %>% 
  select(well, DEPT, x_loc, y_loc, depth_step) %>% 
  ungroup() %>% 
  rename(x= "x_loc", y= "y_loc")


temp<- for_input %>% 
  left_join(help_coord_imput, by= c("well", "DEPT")) %>% 
  group_by(well) %>% 
  mutate(x = min(x, na.rm = T) ,
         y = min(y, na.rm = T),
         x_loc = ifelse(is.na(x_loc), x, x_loc),
         y_loc = ifelse(is.na(y_loc), y, y_loc),
         count = cumsum(ifelse(is.na(z_loc), 1, 0)),
         depth_step = min(depth_step, na.rm = T),
         z_loc = ifelse(count == 0, z_loc , -(count * depth_step) + min(z_loc, na.rm = T))
         ) %>% 
  ungroup()

temp %>% 
  ggplot(aes(y_loc, DEPT))+
  geom_path(size = 2)+
  geom_path(data = for_input, aes(y_loc, DEPT), color = "yellow", size = 0.7)+
  facet_wrap(vars(well), scales = "free")+
  scale_y_reverse()
```

```{r}
#| label: coord_y_plot
#| echo: false

#| fig-width: 8
#| fig-height: 8
#| warning: false

coord_y_imput_plot <-
  temp %>%
  ggplot(aes(y_loc, DEPT)) +
  geom_path(size = 2) +
  geom_path(data = for_input,
            aes(y_loc, DEPT),
            color = "red",
            size = 0.7) +
  facet_wrap(vars(well), scales = "free") +
  scale_y_reverse()+
  theme_minimal()

coord_y_imput_plot

# write_rds(coord_y_imput_plot, "../data/coord_y_imput.rds")
```

```{r}
#| label: fig-coord_y_example
#| echo: false
#| fig-width: 8
#| fig-height: 8
#| warning: false


coord_y_example <-
  temp %>%filter(well%in% "33/9-17") %>% 
  ggplot(aes(y_loc, DEPT)) +
  geom_path(size = 2) +
  geom_path(data = for_input %>% filter(well%in% "33/9-17"),
            aes(y_loc, DEPT),
            color = "yellow",
            size = 0.7) +
  facet_wrap(vars(well), scales = "free") +
  scale_y_reverse()+
  scale_x_continuous(limits = c(6814125,6814225 ))+
  theme_minimal()

coord_y_example

# write_rds(coord_y_example, "../data/coord_y_example.rds")
```

```{r}
#| label: coord_x_plot
#| echo: false
#| fig-width: 8
#| fig-height: 8
#| warning: false

coord_x_imput_plot <-
  temp %>%
  ggplot(aes(x_loc, DEPT)) +
  geom_path(size = 2) +
  geom_path(data = for_input,
            aes(x_loc, DEPT),
            color = "red",
            size = 0.7) +
  facet_wrap(vars(well), scales = "free") +
  scale_y_reverse()+
  theme_minimal()

coord_x_imput_plot

# write_rds(coord_x_imput_plot, "../data/coord_x_imput.rds")
```

```{r}
#| label: fig-un_pozox
#| echo: false
#| fig-width: 8
#| fig-height: 8
#| warning: false


coord_x_example <-
  temp %>%filter(well%in% "33/9-17") %>% 
  ggplot(aes(x_loc, DEPT)) +
  geom_path(size = 2) +
  geom_path(data = for_input %>% filter(well%in% "33/9-17"),
            aes(x_loc, DEPT),
            color = "yellow",
            size = 0.7) +
  facet_wrap(vars(well), scales = "free") +
  scale_y_reverse()+
  scale_x_continuous(limits = c(438400,438500 ))+
  theme_minimal()

coord_x_example

# write_rds(coord_x_example, "../data/coord_x_example.rds")
```

```{r}
#| label: coord_z_plot
#| echo: false
#| fig-width: 8
#| fig-height: 8
#| warning: false

coord_z_imput_plot <-
  temp %>%
  ggplot(aes(-z_loc, DEPT)) +
  geom_path(size = 2) +
  geom_path(
    data = for_input,
    aes(-z_loc, DEPT),
    color = "yellow",
    size = 0.7
  ) +
  facet_wrap(vars(well), scales = "free") +
  scale_y_reverse()+
  theme_minimal()

coord_z_imput_plot

# write_rds(coord_z_imput_plot, "../data/coord_z_imput.rds")
```

```{r}
#| label: coord_z_example
#| echo: false
#| fig-width: 8
#| fig-height: 8
#| warning: false

coord_z_example <-
  temp %>% filter(well%in% "33/9-17") %>% 
  ggplot(aes(-z_loc, DEPT)) +
  geom_path(size = 2) +
  geom_path(
    data = for_input %>% filter(well%in% "33/9-17") ,
    aes(-z_loc, DEPT),
    color = "yellow",
    size = 0.7
  ) +
  facet_wrap(vars(well), scales = "free") +
  scale_y_reverse()+
  theme_minimal()+
  labs(x= "z_loc")

coord_z_example

# write_rds(coord_z_example, "../data/coord_z_example.rds")
```

```{r}
#| label: imputacion
#| echo: false
#| include: false


for_input <-temp %>% select(-c("x", "y", "depth_step", "count"))

no_model_data <- for_input %>% select(well, fm, brent, lith_name, bad_data, x_loc, y_loc, z_loc)

imput_rec <- recipe(DEPT ~ . , data = for_input %>% 
                      select(-c(well, fm, brent, lith_name, bad_data, x_loc, y_loc,z_loc)) )

# imputed_knn <-
#   imput_rec %>%
#   step_impute_knn( all_predictors(),
#                   # columns = "Cholesterol"
#                   ) %>%
#   prep(for_input) %>%
#   bake(for_input) %>%
#   cbind(no_model_data)
# 
# imputed_bag <-
#   imput_rec %>%
#   step_impute_bag(all_predictors(),
#                   # columns = "Cholesterol"
#                   ) %>%
#   prep(for_input) %>%
#   bake(for_input) %>% cbind(no_model_data)
# 
# imputed_lin <-
#   imput_rec %>%
#   step_impute_linear(list_logs_to_impute$Variable_Name,
#                      impute_with = imp_vars(DEPT)
#                   # columns = "Cholesterol"
#                   ) %>%
#   prep(for_input) %>%
#   bake(for_input) %>%  cbind(no_model_data)
# 
# imputed_mean <-
#   imput_rec %>%
#   step_impute_mean(list_logs_to_impute$Variable_Name
#                   # columns = "Cholesterol"
#                   ) %>%
#   prep(for_input) %>%
#   bake(for_input) %>%  cbind(no_model_data)
# 
# imputed_median <-
#   imput_rec %>%
#   step_impute_median(list_logs_to_impute$Variable_Name
#                   # columns = "Cholesterol"
#                   ) %>%
#   prep(for_input) %>%
#   bake(for_input) %>%
#   cbind(no_model_data)
# 
# 
# 
# 
# 
# write_rds(imputed_knn, "../data/imputed_knn.rds")
# write_rds(imputed_bag, "../data/imputed_bag.rds")
# write_rds(imputed_lin, "../data/imputed_lin.rds")
# write_rds(imputed_mean, "../data/imputed_mean.rds")
# write_rds(imputed_median, "../data/imputed_median.rds")

imputed_knn  <-    read_rds( "../data/imputed_knn.rds")
imputed_bag  <-    read_rds( "../data/imputed_bag.rds")
imputed_lin  <-    read_rds( "../data/imputed_lin.rds")
imputed_mean  <-   read_rds( "../data/imputed_mean.rds")
imputed_median  <- read_rds( "../data/imputed_median.rds")

```

```{r}
#| label: fig-juntar_imp
#| echo: false



# all_imputed_orig <-
  # imputed_knn %>%
  # select(c(list_logs_to_impute$Variable_Name, "well")) %>%
  # mutate(imp = "knn") %>%
  # bind_rows(temp %>%
  #             # select(c(list_logs_to_impute$Variable_Name, "well")) %>%
  #             mutate(imp = "orig")) %>%
  # bind_rows(imputed_bag    %>%
  #             # select(c(list_logs_to_impute$Variable_Name, "well")) %>%
  #             mutate(imp = "bag")) %>%
  # bind_rows(imputed_lin    %>%
  #             # select(c(list_logs_to_impute$Variable_Name, "well")) %>%
  #             mutate(imp = "lin")) %>%
  # bind_rows(imputed_mean   %>%
  #             # select(c(list_logs_to_impute$Variable_Name, "well")) %>%
  #             mutate(imp = "mean")) %>%
  # bind_rows(imputed_median %>%
              # # select(c(list_logs_to_impute$Variable_Name, "well")) %>%
              # mutate(imp = "median"))

# all_imputed <- all_imputed_orig %>%
#   select(c(list_logs_to_impute$Variable_Name, "well", "imp")) %>%
#   pivot_longer(-c(well, imp),
#                values_to = "value",
#                names_to = "data")


# write_rds(all_imputed,      "../data/all_imputed.rds"      )
# write_rds(all_imputed_orig, "../data/all_imputed_orig.rds" )
# 
all_imputed      <- read_rds("../data/all_imputed.rds"     )
all_imputed_orig <- read_rds("../data/all_imputed_orig.rds")



all_imputed %>%
  ggplot(aes(value, color = imp)) +
  geom_density() +
  facet_wrap(vars(data), scales = "free", nrow = 3) +
  theme_minimal()
```

```{r}
#| label: all_imput_plot
#| echo: false
#| fig-width: 8
#| fig-height: 8
#| warning: false
#| message: false


all_imputed %>%
  filter(imp %in% c("bag", "knn", "mean", "orig")) %>% 
  ggplot(aes(value, color=imp))+
  geom_density()+
  facet_wrap(vars(data), scales = "free",nrow = 3)+
  theme_minimal()
```

```{r}
#| label: rhob_plot
#| echo: false
#| warning: false
#| message: false



rhob_impute_plot <- 
  all_imputed %>%
  filter(data == "rhob_clean",
         imp %in% c("bag", "knn", "orig")) %>% 
    ggplot(aes(value, color=imp))+
    geom_density()+
    facet_wrap(vars(data), scales = "free",nrow = 2)+
  theme_bw()

rhob_impute_plot
# saveRDS(rhob_impute_plot , "../data/rhob_impute_plot.rds")

```

```{r}
#| label: rhob_table
#| echo: false
#| warning: false
#| message: false


rhob_impute_table <- 
all_imputed %>% filter(data == "rhob_clean", imp %in% c("bag", "knn", "orig", "mean")) %>% 
  group_by(imp) %>% 
  summarise(mean     = round(mean(value, na.rm=TRUE),2),
            # median   = round(median(value, na.rm = TRUE),1),
            sd       = round(sd(value, na.rm = TRUE),2),
            kurtosis = round(kurtosis(value, na.rm = TRUE),2),
             bias =     round(skewness(value, na.rm = TRUE),2 )
            ) %>% 
  flextable() %>% 
  fontsize(size = 8,part = "all") %>% 
  bold(part = "header") %>% 
  bold(i=4, part = "body") %>% 
  color(i=4, color = "red", part = "body") %>% 
  fontsize(i=4, size = 10,part = "body") %>%
  color( i=2,  color = "darkgreen", part = "body") %>% 
  bold(i=2,   part = "body") %>% 
  fontsize(i=2,size = 10,part = "body") %>% 
  add_header_row(values = "RHOB Impute statistics",colwidths = 5)

rhob_impute_table

# saveRDS(rhob_impute_table, "../data/rhob_impute_table.rds")
```

```{r}
#| label: drho_plot
#| echo: false
#| #| warning: false
#| message: false

drhob_impute_plot<-
  all_imputed %>%
  filter(data == "DRHO",
         imp %in% c("bag", "knn", "orig", "mean")) %>% 
  ggplot(aes(value, color=imp))+
  geom_density()+
  facet_wrap(vars(data), scales = "free",nrow = 2)+
  theme_bw()

drhob_impute_plot

# write_rds(drhob_impute_plot, "../data/drhob_impute_plot.rds")
```

```{r}
#| label: drho_table
#| echo: false
#| #| warning: false
#| message: false

drho_impute_table <- 
all_imputed %>% filter(data == "DRHO", imp %in% c("bag", "knn", "orig", "mean")) %>% 
  group_by(imp) %>% 
  summarise(mean     = round(mean(value, na.rm=TRUE),2),
            # median   = round(median(value, na.rm = TRUE),1),
            sd       = round(sd(value, na.rm = TRUE),2),
            kurtosis = round(kurtosis(value, na.rm = TRUE),2),
             bias =     round(skewness(value, na.rm = TRUE),2 )
            ) %>% 
  flextable() %>% 
  fontsize(size = 8,part = "all") %>% 
  bold(part = "header") %>% 
  bold(i=4, part = "body") %>% 
  color(i=4, color = "red", part = "body") %>% 
  fontsize(i=4, size = 10,part = "body") %>%
  color( i=1,  color = "darkgreen", part = "body") %>% 
  bold(i=1,   part = "body") %>% 
  fontsize(i=1,size = 10,part = "body") %>% 
  add_header_row(values = "DRHO Impute statistics",colwidths = 5)

drho_impute_table

# write_rds(drho_impute_table, "../data/drho_impute_table.rds")
```

De los graficos y tabla seleccionamos el mean para DRHO

```{r}
# write_rds(all_imputed, "../data/all_imputed.rds")
# 
# all_imputed <- read_rds("../data/all_imputed.rds")
```

Probamos el PEF

```{r}
#| label: PEF_plot
#| echo: false
#| warning: false
#| message: false

pef_imp_plot <- 
all_imputed %>%
  filter(data == "pef_clean",
         imp %in% c("bag", "knn", "orig", "mean" )) %>%
  ggplot(aes(value, color = imp)) +
  geom_density() +
  facet_wrap(vars(data), scales = "free", nrow = 2) +
  theme_bw()

pef_imp_plot
# write_rds(pef_imp_plot, "../data/pef_imp_plot.rds")

```

revisamos la table del PEF

```{r}
#| label: PEF_table
#| echo: false
#| warning: false
#| message: false


pef_impute_table <- 
all_imputed %>% filter(data == "pef_clean", imp %in% c("bag", "knn", "orig", "mean")) %>% 
  group_by(imp) %>% 
  summarise(mean     = round(mean(value, na.rm=TRUE),2),
            # median   = round(median(value, na.rm = TRUE),1),
            sd       = round(sd(value, na.rm = TRUE),2),
            kurtosis = round(kurtosis(value, na.rm = TRUE),2),
             bias =     round(skewness(value, na.rm = TRUE),2 )
            ) %>% 
  flextable() %>% 
  fontsize(size = 8,part = "all") %>% 
  bold(part = "header") %>% 
  bold(i=4, part = "body") %>% 
  color(i=4, color = "red", part = "body") %>% 
  fontsize(i=4, size = 10,part = "body") %>%
  color( i=1,  color = "darkgreen", part = "body") %>% 
  bold(i=1,   part = "body") %>% 
  fontsize(i=1,size = 10,part = "body") %>% 
  add_header_row(values = "PEF Impute statistics",colwidths = 5)

pef_impute_table

write_rds(pef_impute_table, "../data/pef_impute_table.rds")
```

PEF with bag

```{r}
#| label: SP_plot
#| warning: false
#| echo: false
#| message: false


sp_imp_plot <- 
  all_imputed %>%
  filter(data == "SP",
         imp %in% c("bag", "knn","orig")) %>% 
  ggplot(aes(value, color=imp))+
  geom_density()+
  facet_wrap(vars(data), scales = "free",nrow = 2)+
  theme_bw()

sp_imp_plot

# write_rds(sp_imp_plot, "../data/sp_imp_plot.rds")
```

```{r}
#| label: SP_table
#| echo: false
#| warning: false
#| message: false


sp_impute_table <- 
all_imputed %>% filter(data == "SP", imp %in% c("bag", "knn", "orig", "mean")) %>% 
  group_by(imp) %>% 
  summarise(mean     = round(mean(value, na.rm=TRUE),2),
            # median   = round(median(value, na.rm = TRUE),1),
            sd       = round(sd(value, na.rm = TRUE),2),
            kurtosis = round(kurtosis(value, na.rm = TRUE),2),
             bias =     round(skewness(value, na.rm = TRUE),2 )
            ) %>% 
  flextable() %>% 
  fontsize(size = 8,part = "all") %>% 
  bold(part = "header") %>% 
  bold(i=4, part = "body") %>% 
  color(i=4, color = "red", part = "body") %>% 
  fontsize(i=4, size = 10,part = "body") %>%
  color( i=1,  color = "darkgreen", part = "body") %>% 
  bold(i=1,   part = "body") %>% 
  fontsize(i=1,size = 10,part = "body") %>% 
  add_header_row(values = "SP Impute statistics",colwidths = 5)

sp_impute_table

# write_rds(sp_impute_table, "../data/sp_impute_table.rds")
  
```

### CALI

```{r}
#| label: cali_plot
#| echo: false
#| warning: false
#| message: false


cal_imp_plot <- 
  all_imputed %>%
  filter(data == "CALI",
         imp %in% c("bag", "knn","orig", "mean")) %>% 
  ggplot(aes(value, color=imp))+
  geom_density()+
  facet_wrap(vars(data), scales = "free",nrow = 2)+
  theme_bw()

cal_imp_plot

# write_rds(cal_imp_plot, "../data/cal_imp_plot.rds")
```

```{r}
#| label: cali_table
#| echo: false
#| warning: false
#| message: false

cal_impute_table <- 
all_imputed %>% filter(data == "CALI", imp %in% c("bag", "knn", "orig", "mean")) %>% 
  group_by(imp) %>% 
  summarise(mean     = round(mean(value, na.rm=TRUE),2),
            # median   = round(median(value, na.rm = TRUE),1),
            sd       = round(sd(value, na.rm = TRUE),2),
            kurtosis = round(kurtosis(value, na.rm = TRUE),2),
             bias =     round(skewness(value, na.rm = TRUE),2 )
            ) %>% 
  flextable() %>% 
  fontsize(size = 8,part = "all") %>% 
  bold(part = "header") %>% 
  bold(i=4, part = "body") %>% 
  color(i=4, color = "red", part = "body") %>% 
  fontsize(i=4, size = 10,part = "body") %>%
  color( i=2,  color = "darkgreen", part = "body") %>% 
  bold(i=2,   part = "body") %>% 
  fontsize(i=2,size = 10,part = "body") %>% 
  add_header_row(values = "CALI Impute statistics",colwidths = 5)

cal_impute_table

write_rds(cal_impute_table, "../data/cal_impute_table.rds")
```

we selected the knn

```{r}
#| label: rdep_plot
#| echo: false
#| warning: false
#| message: false


rdep_imp_plot <- 
  all_imputed %>%
  filter(data == "RDEP",
         imp %in% c("bag", "knn","orig", "mean")) %>% 
  ggplot(aes(value, color=imp))+
  geom_density()+
  facet_wrap(vars(data), scales = "free",nrow = 2)+
  theme_bw()+
  scale_x_log10()

rdep_imp_plot

# write_rds(rdep_imp_plot, "../data/rdep_imp_plot.rds")
```

```{r}
#| label: rdep_table
#| echo: false
#| warning: false
#| message: false

rdep_impute_table <- 
all_imputed %>% filter(data == "RDEP", imp %in% c("bag", "knn", "orig", "mean")) %>% 
  group_by(imp) %>% 
  summarise(mean     = round(mean(value, na.rm=TRUE),2),
            # median   = round(median(value, na.rm = TRUE),1),
            sd       = round(sd(value, na.rm = TRUE),2),
            kurtosis = round(kurtosis(value, na.rm = TRUE),2),
             bias =     round(skewness(value, na.rm = TRUE),2 )
            ) %>% 
  flextable() %>% 
  fontsize(size = 8,part = "all") %>% 
  bold(part = "header") %>% 
  bold(i=4, part = "body") %>% 
  color(i=4, color = "red", part = "body") %>% 
  fontsize(i=4, size = 10,part = "body") %>%
  color( i=2,  color = "darkgreen", part = "body") %>% 
  bold(i=2,   part = "body") %>% 
  fontsize(i=2,size = 10,part = "body") %>% 
  add_header_row(values = "RDEP Impute statistics",colwidths = 5)

rdep_impute_table

# write_rds(rdep_impute_table, "../data/rdep_impute_table.rds")
```

now

```{r}
#| label: rsha_plot
#| echo: false
#| warning: false
#| message: false

rsha_imp_plot <- 
  all_imputed %>%
  filter(data == "RSHA",
         imp %in% c("bag", "knn","orig", "mean")) %>% 
  ggplot(aes(value, color=imp))+
  geom_density()+
  facet_wrap(vars(data), scales = "free",nrow = 2)+
  theme_bw()+
  scale_x_log10()

rsha_imp_plot

# write_rds(rsha_imp_plot, "../data/rsha_imp_plot.rds")
```

```{r}
#| label: rsha_table
#| echo: false
#| warning: false
#| message: false


rsha_impute_table <-
all_imputed %>% filter(data == "RSHA", imp %in% c("bag", "knn", "orig", "mean")) %>%
  group_by(imp) %>%
  summarise(mean     = round(mean(value, na.rm=TRUE),2),
            # median   = round(median(value, na.rm = TRUE),1),
            sd       = round(sd(value, na.rm = TRUE),2),
            kurtosis = round(kurtosis(value, na.rm = TRUE),2),
             bias =     round(skewness(value, na.rm = TRUE),2 )
            ) %>%
  flextable() %>%
  fontsize(size = 8,part = "all") %>%
  bold(part = "header") %>%
  bold(i=4, part = "body") %>%
  color(i=4, color = "red", part = "body") %>%
  fontsize(i=4, size = 10,part = "body") %>%
  color( i=2,  color = "darkgreen", part = "body") %>%
  bold(i=2,   part = "body") %>%
  fontsize(i=2,size = 10,part = "body") %>%
  add_header_row(values = "RSHA Impute statistics",colwidths = 5)

rsha_impute_table

# write_rds(rsha_impute_table, "../data/rsha_impute_table.rds")
```

now the RMED

```{r}
#| label: rmed_plot
#| echo: false
#| warning: false
#| message: false


rmed_imp_plot <- 
  all_imputed %>%
  filter(data == "RMED",
         imp %in% c("bag", "knn","orig", "mean")) %>% 
  ggplot(aes(value, color=imp))+
  geom_density()+
  facet_wrap(vars(data), scales = "free",nrow = 2)+
  theme_bw()+
  scale_x_log10()

rmed_imp_plot

# write_rds(rmed_imp_plot, "../data/rmed_imp_plot.rds")
```

```{r}
#| label: rmed_table
#| echo: false
#| warning: false
#| message: false


rmed_impute_table <-
all_imputed %>% filter(data == "RMED", imp %in% c("bag", "knn", "orig", "mean")) %>%
  group_by(imp) %>%
  summarise(mean     = round(mean(value, na.rm=TRUE),2),
            # median   = round(median(value, na.rm = TRUE),1),
            sd       = round(sd(value, na.rm = TRUE),2),
            kurtosis = round(kurtosis(value, na.rm = TRUE),2),
             bias =     round(skewness(value, na.rm = TRUE),2 )
            ) %>%
  flextable() %>%
  fontsize(size = 8,part = "all") %>%
  bold(part = "header") %>%
  bold(i=4, part = "body") %>%
  color(i=4, color = "red", part = "body") %>%
  fontsize(i=4, size = 10,part = "body") %>%
  color( i=1,  color = "darkgreen", part = "body") %>%
  bold(i=1,   part = "body") %>%
  fontsize(i=1,size = 10,part = "body") %>%
  add_header_row(values = "RMED Impute statistics",colwidths = 5)

rmed_impute_table

# write_rds(rmed_impute_table, "../data/rmed_impute_table.rds")
```

```{r}
#| label: dtc_plot
#| echo: false
#| warning: false
#| message: false


dtc_imp_plot <- 
  all_imputed %>%
  filter(data == "DTC",
         imp %in% c("bag", "knn","orig", "mean")) %>% 
  ggplot(aes(value, color=imp))+
  geom_density()+
  facet_wrap(vars(data), scales = "free",nrow = 2)+
  theme_bw()+
  scale_x_log10()

dtc_imp_plot

# write_rds(dtc_imp_plot, "../data/dtc_imp_plot.rds")
```

```{r}
#| label: dtc_table
#| echo: false
#| warning: false
#| message: false


dtc_impute_table <-
all_imputed %>% filter(data == "DTC", imp %in% c("bag", "knn", "orig", "mean")) %>%
  group_by(imp) %>%
  summarise(mean     = round(mean(value, na.rm=TRUE),2),
            # median   = round(median(value, na.rm = TRUE),1),
            sd       = round(sd(value, na.rm = TRUE),2),
            kurtosis = round(kurtosis(value, na.rm = TRUE),2),
             bias =     round(skewness(value, na.rm = TRUE),2 )
            ) %>%
  flextable() %>%
  fontsize(size = 8,part = "all") %>%
  bold(part = "header") %>%
  bold(i=4, part = "body") %>%
  color(i=4, color = "red", part = "body") %>%
  fontsize(i=4, size = 10,part = "body") %>%
  color( i=2,  color = "darkgreen", part = "body") %>%
  bold(i=2,   part = "body") %>%
  fontsize(i=2,size = 10,part = "body") %>%
  add_header_row(values = "DTC Impute statistics",colwidths = 5)

dtc_impute_table

# write_rds(dtc_impute_table, "../data/dtc_impute_table.rds")

```

```{r}
#| label: nphi_plot
#| echo: false
#| warning: false
#| message: false


nphi_imp_plot <- 
  all_imputed %>%
  filter(data == "NPHI",
         imp %in% c("bag", "knn","orig", "mean")) %>% 
  ggplot(aes(value, color=imp))+
  geom_density()+
  facet_wrap(vars(data), scales = "free",nrow = 2)+
  theme_bw()+
  scale_x_log10()

nphi_imp_plot

# write_rds(nphi_imp_plot, "../data/nphi_imp_plot.rds")
```

```{r}
#| label: nphi_table
#| echo: false
#| warning: false
#| message: false


nphi_impute_table <-
all_imputed %>% filter(data == "NPHI", imp %in% c("bag", "knn", "orig", "mean")) %>%
  group_by(imp) %>%
  summarise(mean     = round(mean(value, na.rm=TRUE),2),
            # median   = round(median(value, na.rm = TRUE),1),
            sd       = round(sd(value, na.rm = TRUE),2),
            kurtosis = round(kurtosis(value, na.rm = TRUE),2),
             bias =     round(skewness(value, na.rm = TRUE),2 )
            ) %>%
  flextable() %>%
  fontsize(size = 8,part = "all") %>%
  bold(part = "header") %>%
  bold(i=4, part = "body") %>%
  color(i=4, color = "red", part = "body") %>%
  fontsize(i=4, size = 10,part = "body") %>%
  color( i=2,  color = "darkgreen", part = "body") %>%
  bold(i=2,   part = "body") %>%
  fontsize(i=2,size = 10,part = "body") %>%
  add_header_row(values = "NPHI Impute statistics",colwidths = 5)

nphi_impute_table

# write_rds(nphi_impute_table, "../data/nphi_impute_table.rds")
```

### GR

```{r}
#| label: fig-gr_plot
#| echo: false
#| warning: false
#| message: false


gr_imp_plot <- 
  all_imputed %>%
  filter(data == "GR",
         imp %in% c("bag", "knn","orig", "mean")) %>% 
  ggplot(aes(value, color=imp))+
  geom_density()+
  facet_wrap(vars(data), scales = "free",nrow = 2)+
  theme_bw()+
  scale_x_log10()

gr_imp_plot

# write_rds(gr_imp_plot, "../data/gr_imp_plot.rds")
```

```{r}
#| label: gr_table
#| echo: false
#| warning: false
#| message: false


gr_impute_table <-
all_imputed %>% filter(data == "GR", imp %in% c("bag", "knn", "orig", "mean")) %>%
  group_by(imp) %>%
  summarise(mean     = round(mean(value, na.rm=TRUE),2),
            # median   = round(median(value, na.rm = TRUE),1),
            sd       = round(sd(value, na.rm = TRUE),2),
            kurtosis = round(kurtosis(value, na.rm = TRUE),2),
             bias =     round(skewness(value, na.rm = TRUE),2 )
            ) %>%
  flextable() %>%
  fontsize(size = 8,part = "all") %>%
  bold(part = "header") %>%
  bold(i=4, part = "body") %>%
  color(i=4, color = "red", part = "body") %>%
  fontsize(i=4, size = 10,part = "body") %>%
  color( i=3,  color = "darkgreen", part = "body") %>%
  bold(i=3,   part = "body") %>%
  fontsize(i=3,size = 10,part = "body") %>%
  add_header_row(values = "NPHI Impute statistics",colwidths = 5)

gr_impute_table

# write_rds(gr_impute_table, "../data/gr_impute_table.rds")
```

## Imputaciones utilizadas

```{r}
#| label: imputaciones_table
#| echo: false
#| warning: false
#| message: false

imput_algo <- tibble(Curva = c("CALI", "RDEP", "RSHA", "RMED", "SP", "DTC", "NPHI", "DRHO", "rhob_clean", "pef_clean" , "GR"),
                     imput = c("knn",   "knn", "knn",  "bag",  "bag", "knn", "knn", "bag", "knn",         "bag"  , "mean")) 

# write_rds(imput_algo, "../data/imput_algo.rds")
  

```

Visualización de las imputaciones

```{r}
#| echo: false
#| warning: false
#| message: false
#| 
imputed_data <-
  imputed_bag %>% 
  select(well, DEPT, x_loc, y_loc, z_loc, RMED, SP, DRHO, pef_clean, bit_size,lith_name, fm) %>%
  cbind(imputed_mean %>%
          select(GR)) %>%
  cbind(imputed_knn %>%
          select(CALI, RDEP, RSHA, DTC, NPHI, rhob_clean))

# write_rds(imputed_data, "../data/imputed_data.rds")

```
